Move States into separate subfolder
