+++ src/Stores/States/EggState.js
